# [commands] faster sqldump query scanner