--- a/go/cmd/dolt/commands/sql_statement_scanner_test.go
+++ b/go/cmd/dolt/commands/sql_statement_scanner_test.go
@@ -185,6 +185,7 @@ insert into foo values (1,2,3)|`,
 				"",
 				"insert into foo values (1,2,3)",
 			},
+			lineNums: []int{1, 2},
 		},
 	}